FEP-3264 Federated Work Coordination (#244)
This is one of two "planning and doing" FEPs.

I don't know if this is too big or not, still thinking through the best way to extend AP with Valueflows.  This is meant to continue @silverpill 's use-case level approach in FEP-0837 Federated Marketplace.

I look forward to getting feedback on that question.  And anything else, of course.

Reviewed-on: https://codeberg.org/fediverse/fep/pulls/244
Co-authored-by: Lynn Foster <lynn@mikorizal.org>
Co-committed-by: Lynn Foster <lynn@mikorizal.org>