FEP 6841: Specifying ActivityPub extension support with NodeInfo (#278)
Motivated by the need to know if a remote server supports a particular ActivityPub extension, I've knocked this proposed solution together based on defining extension IRIs, and adding them into a server's NodeInfo response.

I'm totally open to changing pretty much anything about this, it's as much a prompt for discussion as a completely thought-out proposal.

Reviewed-on: https://codeberg.org/fediverse/fep/pulls/278
Co-authored-by: James Smith <james@floppy.org.uk>
Co-committed-by: James Smith <james@floppy.org.uk>