[FEP-521a] Add human-readable testing script (#342)
Sorry, it's a lot more verbose and esoteric than a gherkin, but that's what's we're using! I figured it could maybe live here canonically as per !328, while we're working on the rest of the suite.  In the meantime, I can put a little free-standing single-test-case-runner link in the tracking issue, as you mentioned.

Co-authored-by: bumblefudge <bumblefudge@learningproof.xyz>
Reviewed-on: https://codeberg.org/fediverse/fep/pulls/342
Co-authored-by: bumblefudge <bumblefudge@noreply.codeberg.org>
Co-committed-by: bumblefudge <bumblefudge@noreply.codeberg.org>