[FEP-0f2a]: Announce Activity for Migrations and Tombstone Events (#351)
This thread is mostly SHOULDs and pedantically-explicit explanations of how existing FEPs and existing AP/AS behaviors work together.  Buried among the self-evident stuff are a _few_ novel behaviors and ideas, but the _main_ purpose was to make existing FEPs easier to test (see the attached specification for a Test Case, with sample implementation forthcoming).

Co-authored-by: bumblefudge <bumblefudge@learningproof.xyz>
Co-authored-by: Juan Caballero <juan@webarch.coop>
Co-authored-by: bengo <171782+gobengo@users.noreply.github.com>
Reviewed-on: https://codeberg.org/fediverse/fep/pulls/351
Co-authored-by: bumblefudge <bumblefudge@noreply.codeberg.org>
Co-committed-by: bumblefudge <bumblefudge@noreply.codeberg.org>