FEP-eb22: Supported ActivityStreams types with NodeInfo (#378)
Servers can advertise what features of the API they support, such as creating a poll or boosting a post. Clients can recognize if a server doesn't support a feature and hide it from the UI.

This proposal overlaps a little with FEP-9fde and FEP-6481. Special thanks to Nik Clayton who inspired this with his FEP and discussion at FediForum. The main difference with this is we attempt to describe what features a server supports based on ActivityStreams types instead of new identifiers or URIs. For that reason, it felt right to start with a new FEP rather than update either of those existing FEPs.

I'm happy to collaborate with anyone on this. When there is some consensus that we're on the right track, I plan to take an open source Mastodon API client and adapt it to support this as a proof of concept.

Reviewed-on: https://codeberg.org/fediverse/fep/pulls/378
Co-authored-by: Manton Reece <manton@me.com>
Co-committed-by: Manton Reece <manton@me.com>