FEP-a4ed: Replace the ASCII diagram with Mermaid (#394)
This patch replaces the ASCII diagram with a [Mermaid] diagram in order to improve accessibility for screen readers/braille displays and devices with narrow viewport width.

[Rendered](https://codeberg.org/tesaguri/fep/src/branch/fep-a4ed-mermaid/fep/a4ed/fep-a4ed.md#the-fediverse-enhancement-proposals-process)

While rendering Mermaid diagram isn't part of CommonMark 0.29 spec used by the FEP process, this uses the standard code fence syntax and isn't extremely unreadable as plaintext, so I believe it's acceptable in that regard.

[Mermaid]: <https://docs.codeberg.org/markdown/introduction-to-markdown/#mermaid-diagrams>

Reviewed-on: https://codeberg.org/fediverse/fep/pulls/394
Reviewed-by: silverpill <silverpill@noreply.codeberg.org>
Reviewed-by: Arnold Schrijver <circlebuilder@noreply.codeberg.org>
Reviewed-by: helge <helge@noreply.codeberg.org>
Co-authored-by: Daiki Mizukami <tesaguriguma@gmail.com>
Co-committed-by: Daiki Mizukami <tesaguriguma@gmail.com>