FEP-044f: Consent-respecting quote posts (#545)
This is a work-in-progress document describing Mastodon's proposed way of representing quote posts, users' choices regarding whether their posts can be quoted and by whom (quote policies), and a mechanism for servers to verify compliance with such policies.

This document describes protocol considerations, which do not necessarily translate directly to User Experience considerations. For instance, the use of the approval mechanism described in this document does not imply that the user's approval is manual.

Co-authored-by: Claire <claire.github-309c@sitedethib.com>
Reviewed-on: https://codeberg.org/fediverse/fep/pulls/545
Co-authored-by: ClearlyClaire <clearlyclaire@noreply.codeberg.org>
Co-committed-by: ClearlyClaire <clearlyclaire@noreply.codeberg.org>