Add FEP-82f6: Actor statuses (#596)
This proposal describes an ActivityPub extension to allow actors to publish a short status text, with optional expiration, link attachment, and history.

Some centralized communication services provide their users with the ability to set a status on their account, which is usually displayed on their profile and sometimes next to their name in other places in the UI. These are distinct from regular posts because they can not be interacted with in any way whatsoever, can't contain media attachments, and usually have a short character limit on the order of several hundred characters at most. Statuses are always visible to anyone who can see the actor itself.

Co-authored-by: Grishka <git@grishka.me>
Reviewed-on: https://codeberg.org/fediverse/fep/pulls/596
Co-authored-by: grishka <grishka@noreply.codeberg.org>
Co-committed-by: grishka <grishka@noreply.codeberg.org>