FEP-a4ed: Withdraw stale proposals (#543)
FEP-a4ed requires facilitators to contact authors of a stale proposal before withdrawing it. As our experience shows, that doesn't work. In this PR I propose withdrawing stale proposals after 1 year of inactivity without contacting authors.

This is a major change. I think even if we agree on it, existing proposals should be handled according to the old procedure.

Resolves https://codeberg.org/fediverse/fep/issues/101

Reviewed-on: https://codeberg.org/fediverse/fep/pulls/543
Reviewed-by: Arnold Schrijver <circlebuilder@noreply.codeberg.org>
Co-authored-by: silverpill <silverpill@firemail.cc>
Co-committed-by: silverpill <silverpill@firemail.cc>